+++ opal/hospital_settings/factories.py
@@ -1,4 +1,5 @@
 """Module providing model factories for hospital settings app models."""
+
 import factory
 from factory.django import DjangoModelFactory
 
@@ -14,6 +15,8 @@ class Institution(DjangoModelFactory):
 
     name = 'McGill University Health Centre'
     name_fr = 'Centre universitaire de santé McGill'
+    logo = factory.django.ImageField(from_path='opal/hospital_settings/tests/fixtures/test_logo.png')
+    logo_fr = factory.django.ImageField(from_path='opal/hospital_settings/tests/fixtures/test_logo.png')
     code = factory.lazy_attribute(lambda institution: institution.name[:4].upper())