+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Add patient model (!98)
 * Feat: Get all data needed for the app home view request (!118)
 * feat: Registration encryption info endpoint (!170)
+* Add logo (image) field to the Institution model (!159)
 
 ### Documentation